--- src/main/java/com/aiad/matcher/JunitPractice2.java
@@ -1,35 +1,0 @@
-package com.aiad.matcher;
-
-import junit.framework.TestCase;
-
-import static org.hamcrest.CoreMatchers.*;
-import static org.hamcrest.Matchers.*;
-import static org.junit.Assert.assertThat;
-
-
-/**
- * 文字列用のMatcher.
- * @author kishibashi
- */
-public class JunitPractice2 extends TestCase{
-	
-	
-	public void testStartsWith(){
-		
-	}
-    public void testEndsWith(){
-    	
-    }
-    public void testContainsString(){
-    	
-    }
-    public void testEqualToIgnoreCase(){
-    	
-    }
-    public void testIsEmptyString(){
-    	
-    }
-
-}
-
-
