+++ src/main/java/com/aiad/practice/junit/matcher/JunitPractice2.java
@@ -1,8 +1,8 @@
-package com.aiad.matcher;
-
-import org.hamcrest.text.IsEqualIgnoringCase;
+package com.aiad.practice.junit.matcher;
 
 import junit.framework.TestCase;
+
+import static org.hamcrest.CoreMatchers.*;
 import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.assertThat;
 
@@ -11,22 +11,23 @@ import static org.junit.Assert.assertThat;
  * 文字列用のMatcher.
  * @author kishibashi
  */
-public class JunitPractice2Test extends TestCase{
+public class JunitPractice2 extends TestCase{
+	
 	
 	public void testStartsWith(){
-		assertThat("aaa", startsWith("b"));
+		
 	}
     public void testEndsWith(){
-    	assertThat("aaa", endsWith("b"));
+    	
     }
     public void testContainsString(){
-    	assertThat("aaa", containsString("b"));
+    	
     }
     public void testEqualToIgnoreCase(){
-    	assertThat("Abc", equalToIgnoringCase("abC"));
+    	
     }
     public void testIsEmptyString(){
-    	assertThat("", isEmptyString());
+    	
     }
 
 }
