+++ src/main/java/com/aiad/practice/junit/mockito/JUnitPractice7.java
@@ -1,4 +1,4 @@
-package com.aiad.mockito;
+package com.aiad.practice.junit.mockito;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
